resume-screening/requirements.txt
streamlit
scikit-learn
pandas
PyPDF2
python-docx
pickle-mixin
